Added variable step ticker class abstracted from WS

+++ wildflower-core/src/main/java/wildflower/World.java
@@ -64,7 +64,7 @@ public class World {
         return this.entities.values();
     }
 
-    public void update(double delta) {
+    public void update(float delta) {
         this.reapDeadCreatures();
         this.updateEntities();
     }
